sage-pro/requirements-core.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0.0
structlog>=24.0.0
sse-starlette>=1.6.0
python-jose[cryptography]>=3.3.0
google-auth>=2.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
langgraph>=0.1.0
faiss-cpu>=1.7.0
python-Levenshtein>=0.21.0
gudhi>=3.8.0
datasets>=2.0.0
hypothesis>=6.0.0
pytest>=7.0.0
radon>=5.1.0
requests>=2.28.0
